Refactor: Update data types and date formatting
- Changed `mileage` in `ShiftV2` from `Double` to `Long`.
- Updated mileage calculation in `ShiftV2.toShift()` to divide by 100.
- Changed data types for `rentCost`, `serviceCost`, `consumption`, and `totalMileage` in `Car` from `Double` to `Long`.
- Modified `DeprecatedDateFormatter` from an object to a top-level `val`.
- Updated `ShiftRepository` and `ShiftV2` to use the new `DeprecatedDateFormatter`.

+++ app/src/main/java/com/hsact/taxilog/data/model/shift2/Car.kt
@@ -3,8 +3,8 @@ package com.hsact.taxilog.data.model.shift2
 data class Car(
     val id: Int,
     val name: String,
-    val rentCost: Double = 0.0,
-    val serviceCost: Double = 0.0,
-    val consumption: Double = 0.0,
-    val totalMileage: Double = 0.0
+    val rentCost: Long = 0,
+    val serviceCost: Long = 0,
+    val consumption: Long = 0,
+    val totalMileage: Long = 0
 )
